python-service/requirements.txt
et-xmlfile==1.1.0
openpyxl==3.1.4
Flask==3.0.3
openai==1.34.0
python-dotenv==1.0.1
